LayoutTests/fast/shadow-dom/manual-slot-gc-crash-expected.txt

PASS slot.assignedNodes().length is 0
PASS successfullyParsed is true

TEST COMPLETE
